feat: DBのAxiosインスタンスを生成する処理を簡略化しました。
