+++ src/utils/getAnime.js
@@ -1,10 +1,10 @@
-import createDbAxios from "./createDbAxios";
+import dbAxios from "./dbAxios";
 
 const getAnime = async () => {
-  const animeAxios = createDbAxios("databases/9ec213bc61bd4af5bbdfc74b844cb4b7/query")
+  const PATH = "databases/9ec213bc61bd4af5bbdfc74b844cb4b7/query";
 
   try {
-    const response = await animeAxios.post();
+    const response = await dbAxios.post(PATH);
     return response.data.results;
   } catch (error) {
     console.error(error);
