+++ src/utils/getReviews.js
@@ -1,10 +1,10 @@
-import createDbAxios from "./createDbAxios";
+import dbAxios from "./dbAxios";
 
 const getReviews = async () => {
-  const reviewsAxios = createDbAxios("databases/d571c5897d854e54bd89275d12338a35/query")
+  const PATH = "databases/d571c5897d854e54bd89275d12338a35/query";
 
   try {
-    const response = await reviewsAxios.post();
+    const response = await dbAxios.post(PATH);
     return response.data.results;
   } catch (error) {
     console.error(error);
